+++ lib/picker.dart
@@ -50,7 +50,12 @@ class MultiImagePicker {
 
     var assets = List<Asset>();
     for (var item in images) {
-      var asset = Asset(item['identifier'], item['width'], item['height']);
+      var asset = Asset(
+        item['identifier'],
+        item['name'],
+        item['width'],
+        item['height'],
+      );
       assets.add(asset);
     }
     return assets;